Move some of the ActiveRecord patches to ApplicationJob
