+++ app/models/application_record.rb
@@ -1,4 +1,7 @@
 # frozen_string_literal: true
 class ApplicationRecord < ActiveRecord::Base
   self.abstract_class = true
+
+  include ApplicationUserstampConcern
+  include ApplicationActsAsConcern
 end
