--- lib/extensions/association_inverse_suppression/active_record.rb
@@ -1,2 +1,0 @@
-# frozen_string_literal: true
-module Extensions::AssociationInverseSuppression::ActiveRecord; end
